Modifying sigmoid saturation
