+++ pymc_marketing/mmm/transformers.py
@@ -749,24 +749,3 @@ def michaelis_menten(
     """
 
     return alpha * x / (lam + x)
-
-
-def sigmoid_saturation(
-    x: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    alpha: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    lam: Union[float, np.ndarray, npt.NDArray[np.float64]],
-) -> Union[float, Any]:
-    """
-    Parameters
-    ----------
-    - alpha
-        α (alpha): Represent the Asymptotic Maximum or Ceiling Value.
-    - lam
-        λ (lambda): affects how quickly the function approaches its upper and lower asymptotes. A higher value of
-        lam makes the curve steeper, while a lower value makes it more gradual.
-    """
-
-    if alpha <= 0 or lam <= 0:
-        raise ValueError("alpha and lam must be greater than 0")
-
-    return (alpha - alpha * np.exp(-lam * x)) / (1 + np.exp(-lam * x))
