+++ tests/mmm/test_transformers.py
@@ -17,7 +17,6 @@ from pymc_marketing.mmm.transformers import (
     logistic_saturation,
     michaelis_menten,
     scale_preserving_logistic_saturation,
-    sigmoid_saturation,
     tanh_saturation,
     tanh_saturation_baselined,
     weibull_adstock,
@@ -448,28 +447,6 @@ class TestSaturationTransformers:
     )
     def test_michaelis_menten(self, x, alpha, lam, expected):
         assert np.isclose(michaelis_menten(x, alpha, lam), expected, atol=0.01)
-
-    @pytest.mark.parametrize(
-        "x, alpha, lam, expected",
-        [
-            (0, 1, 1, 0),
-            (1, 1, 1, 0.4621),
-        ],
-    )
-    def test_sigmoid_saturation(self, x, alpha, lam, expected):
-        assert np.isclose(sigmoid_saturation(x, alpha, lam), expected, atol=0.01)
-
-    @pytest.mark.parametrize(
-        "x, alpha, lam",
-        [
-            (0, 0, 1),
-            (1, -1, 1),
-            (1, 1, 0),
-        ],
-    )
-    def test_sigmoid_saturation_value_errors(self, x, alpha, lam):
-        with pytest.raises(ValueError):
-            sigmoid_saturation(x, alpha, lam)
 
 
 class TestTransformersComposition:
